Add TeamTasks, TeamFinance, UserAchievement repos
Add new repository interfaces and implementations for team tasks (TeamTasksRepository + TeamTasksRepositoryImpl), team finance/transactions & reports (TeamFinanceRepository + TeamFinanceRepositoryImpl), and user achievements (UserAchievementRepository + UserAchievementRepositoryImpl). Refactor TeamsRepository and UserRepository to extend and delegate to these specialized repositories, remove duplicated task/finance/achievement logic from TeamsRepositoryImpl and UserRepositoryImpl, and register the new bindings in RepositoryModule. This splits responsibilities for tasks, finance and achievement flows, improving separation of concerns, modularity and testability.
